PS-2031 Extract TableWriter::uploadTable() into a separate class
