+++ src/Writer/TableWriter.php
@@ -2,11 +2,8 @@
 
 namespace Keboola\OutputMapping\Writer;
 
-use Keboola\Csv\CsvFile;
-use Keboola\Csv\Exception;
 use Keboola\OutputMapping\Configuration\Table\Manifest as TableManifest;
 use Keboola\OutputMapping\Configuration\Table\Manifest\Adapter as TableAdapter;
-use Keboola\OutputMapping\DeferredTasks\LoadTable;
 use Keboola\OutputMapping\DeferredTasks\LoadTableQueue;
 use Keboola\OutputMapping\DeferredTasks\MetadataDefinition;
 use Keboola\OutputMapping\Exception\InvalidOutputException;
@@ -18,11 +15,6 @@ use Keboola\OutputMapping\Writer\Helper\ManifestHelper;
 use Keboola\OutputMapping\Writer\Helper\PrimaryKeyHelper;
 use Keboola\StorageApi\ClientException;
 use Keboola\StorageApi\Metadata;
-use Keboola\StorageApi\Options\FileUploadOptions;
-use Keboola\Temp\Temp;
-use MicrosoftAzure\Storage\Blob\BlobRestProxy;
-use MicrosoftAzure\Storage\Blob\Models\ListBlobsOptions;
-use MicrosoftAzure\Storage\Common\Exceptions\ServiceException;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
@@ -30,32 +22,18 @@ use Symfony\Component\Finder\SplFileInfo;
 
 class TableWriter extends AbstractWriter
 {
-    const SYSTEM_METADATA_PROVIDER = 'system';
-    const KBC_LAST_UPDATED_BY_BRANCH_ID = 'KBC.lastUpdatedBy.branch.id';
-    const KBC_LAST_UPDATED_BY_CONFIGURATION_ROW_ID = 'KBC.lastUpdatedBy.configurationRow.id';
-    const KBC_LAST_UPDATED_BY_CONFIGURATION_ID = 'KBC.lastUpdatedBy.configuration.id';
-    const KBC_LAST_UPDATED_BY_COMPONENT_ID = 'KBC.lastUpdatedBy.component.id';
-    const KBC_CREATED_BY_BRANCH_ID = 'KBC.createdBy.branch.id';
-    const KBC_CREATED_BY_CONFIGURATION_ROW_ID = 'KBC.createdBy.configurationRow.id';
-    const KBC_CREATED_BY_CONFIGURATION_ID = 'KBC.createdBy.configuration.id';
-    const KBC_CREATED_BY_COMPONENT_ID = 'KBC.createdBy.component.id';
-
-    /** @var Metadata */
-    private $metadataClient;
-
-    /** @var string */
-    private $stagingStorageOutput;
-
-    /** @var Table\StrategyInterface */
-    private $strategy;
-
-    /** @var string */
-    private $sourcePath;
+    /** @var TableUploader */
+    private $tableUploader;
 
     public function __construct(StrategyFactory $strategyFactory)
     {
         parent::__construct($strategyFactory);
-        $this->metadataClient = new Metadata($this->clientWrapper->getBasicClient());
+
+        $this->tableUploader = new TableUploader(
+            $this->clientWrapper,
+            new Metadata($this->clientWrapper->getBasicClient()),
+            $this->logger
+        );
     }
 
     /**
@@ -72,14 +50,11 @@ class TableWriter extends AbstractWriter
             throw new OutputOperationException('Component Id must be set');
         }
 
-        $this->stagingStorageOutput = $stagingStorageOutput;
-        $this->strategy = $this->strategyFactory->getTableOutputStrategy($stagingStorageOutput);
-
         if ($stagingStorageOutput === StrategyFactory::LOCAL) {
-            return $this->uploadTablesLocal($source, $configuration, $systemMetadata);
-        } else {
-            return $this->uploadTablesWorkspace($source, $configuration, $systemMetadata);
+            return $this->uploadTablesLocal($source, $configuration, $systemMetadata, $stagingStorageOutput);
         }
+
+        return $this->uploadTablesWorkspace($source, $configuration, $systemMetadata, $stagingStorageOutput);
     }
 
     /**
@@ -89,12 +64,13 @@ class TableWriter extends AbstractWriter
      * @return LoadTableQueue
      * @throws \Exception
      */
-    private function uploadTablesWorkspace($source, array $configuration, array $systemMetadata)
+    private function uploadTablesWorkspace($source, array $configuration, array $systemMetadata, $stagingStorageOutput)
     {
-        $this->sourcePath = $source;
+        $strategy = $this->strategyFactory->getTableOutputStrategy($stagingStorageOutput);
+
         $finder = new Finder();
         /** @var SplFileInfo[] $files */
-        $files = $finder->files()->name('*.manifest')->in($this->strategy->getMetadataStorage()->getPath() . '/' . $source)->depth(0);
+        $files = $finder->files()->name('*.manifest')->in($strategy->getMetadataStorage()->getPath() . '/' . $source)->depth(0);
 
         $sources = [];
         // add output mappings fom configuration
@@ -153,7 +129,14 @@ class TableWriter extends AbstractWriter
                 $parsedConfig['primary_key'] =
                     PrimaryKeyHelper::normalizeKeyArray($this->logger, $parsedConfig['primary_key']);
                 $parsedConfig = DestinationRewriter::rewriteDestination($parsedConfig, $this->clientWrapper);
-                $tableJob = $this->uploadTable($sourceName, $parsedConfig, $systemMetadata);
+                $tableJob = $this->tableUploader->uploadTable(
+                    $strategy,
+                    $stagingStorageOutput,
+                    $source,
+                    $sourceName,
+                    $parsedConfig,
+                    $systemMetadata
+                );
             } catch (ClientException $e) {
                 throw new InvalidOutputException(
                     "Cannot upload file '{$sourceName}' to table '{$parsedConfig["destination"]}' in Storage API: "
@@ -198,15 +181,18 @@ class TableWriter extends AbstractWriter
      * @param string $source
      * @param array $configuration
      * @param array $systemMetadata
+     * @param $stagingStorageOutput
      * @return LoadTableQueue
      * @throws \Exception
      */
-    private function uploadTablesLocal($source, array $configuration, array $systemMetadata)
+    private function uploadTablesLocal($source, array $configuration, array $systemMetadata, $stagingStorageOutput)
     {
+        $strategy = $this->strategyFactory->getTableOutputStrategy($stagingStorageOutput);
+
         if (empty($systemMetadata[self::SYSTEM_KEY_COMPONENT_ID])) {
             throw new OutputOperationException('Component Id must be set');
         }
-        $manifestNames = ManifestHelper::getManifestFiles($this->strategy->getMetadataStorage()->getPath() . '/' . $source);
+        $manifestNames = ManifestHelper::getManifestFiles($strategy->getMetadataStorage()->getPath() . '/' . $source);
 
         $finder = new Finder();
 
@@ -220,7 +206,7 @@ class TableWriter extends AbstractWriter
         $processedOutputMappingTables = [];
 
         /** @var SplFileInfo[] $files */
-        $files = $finder->notName('*.manifest')->in($this->strategy->getDataStorage()->getPath() . '/' . $source)->depth(0);
+        $files = $finder->notName('*.manifest')->in($strategy->getDataStorage()->getPath() . '/' . $source)->depth(0);
 
         $fileNames = [];
         foreach ($files as $file) {
@@ -306,7 +292,14 @@ class TableWriter extends AbstractWriter
             try {
                 $config['primary_key'] = PrimaryKeyHelper::normalizeKeyArray($this->logger, $config['primary_key']);
                 $config = DestinationRewriter::rewriteDestination($config, $this->clientWrapper);
-                $tableJob = $this->uploadTable($file->getPathname(), $config, $systemMetadata);
+                $tableJob = $this->tableUploader->uploadTable(
+                    $strategy,
+                    $stagingStorageOutput,
+                    $source,
+                    $file->getPathname(),
+                    $config,
+                    $systemMetadata
+                );
             } catch (ClientException $e) {
                 throw new InvalidOutputException(
                     "Cannot upload file '{$file->getFilename()}' to table '{$config["destination"]}' in Storage API: "
@@ -349,7 +342,7 @@ class TableWriter extends AbstractWriter
         );
         if (count($diff)) {
             throw new InvalidOutputException(
-                sprintf('Can not process output mapping for file(s): %s.', join('", "', $diff))
+                sprintf('Can not process output mapping for file(s): %s.', implode('", "', $diff))
             );
         }
         $tableQueue = new LoadTableQueue($this->clientWrapper->getBasicClient(), $jobs);
@@ -391,354 +384,10 @@ class TableWriter extends AbstractWriter
     private function createDestinationConfigParam($prefix, $filename)
     {
         if (substr($filename, -4) === '.csv') {
-            return $prefix . substr($filename, 0, strlen($filename) - 4);
-        } else {
-            return $prefix . $filename;
-        }
-    }
-
-    private function checkDevBucketMetadata($destination)
-    {
-        if (!$this->clientWrapper->hasBranch()) {
-            return;
-        }
-        $bucketId = $this->getBucketId($destination);
-        $metadata = new Metadata($this->clientWrapper->getBasicClient());
-        try {
-            foreach ($metadata->listBucketMetadata($bucketId) as $metadatum) {
-                if (($metadatum['key'] === self::KBC_LAST_UPDATED_BY_BRANCH_ID) ||
-                    ($metadatum['key'] === self::KBC_CREATED_BY_BRANCH_ID)) {
-                    if ((string) $metadatum['value'] === (string) $this->clientWrapper->getBranchId()) {
-                        return;
-                    } else {
-                        throw new InvalidOutputException(sprintf(
-                            'Trying to create a table in the development bucket "%s" on branch ' .
-                            '"%s" (ID "%s"). The bucket metadata marks it as assigned to branch with ID "%s".',
-                            $bucketId,
-                            $this->clientWrapper->getBranchName(),
-                            $this->clientWrapper->getBranchId(),
-                            $metadatum['value']
-                        ));
-                    }
-                }
-            }
-        } catch (ClientException $e) {
-            // this is Ok, if the bucket it does not exists, it can't have wrong metadata
-            if ($e->getCode() === 404) {
-                return;
-            } else {
-                throw $e;
-            }
-        }
-        throw new InvalidOutputException(sprintf(
-            'Trying to create a table in the development ' .
-            'bucket "%s" on branch "%s" (ID "%s"), but the bucket is not assigned to any development branch.',
-            $bucketId,
-            $this->clientWrapper->getBranchName(),
-            $this->clientWrapper->getBranchId()
-        ));
-    }
-
-    /**
-     * @param string $source
-     * @param array $config
-     * @param array $systemMetadata
-     * @return LoadTable
-     * @throws ClientException
-     */
-    private function uploadTable($source, array $config, array $systemMetadata)
-    {
-        if (is_dir($source) && empty($config['columns'])) {
-            throw new InvalidOutputException(
-                sprintf('Sliced file "%s" columns specification missing.', basename($source))
-            );
-        }
-        if (!$this->clientWrapper->getBasicClient()->bucketExists($this->getBucketId($config['destination']))) {
-            $this->createBucket($config['destination'], $systemMetadata);
-        } else {
-            $this->checkDevBucketMetadata($config['destination']);
+            return $prefix . substr($filename, 0, -4);
         }
 
-        if ($this->clientWrapper->getBasicClient()->tableExists($config['destination'])) {
-            $tableInfo = $this->clientWrapper->getBasicClient()->getTable($config['destination']);
-            PrimaryKeyHelper::validatePrimaryKeyAgainstTable($this->logger, $tableInfo, $config);
-            if (PrimaryKeyHelper::modifyPrimaryKeyDecider($this->logger, $tableInfo, $config)) {
-                PrimaryKeyHelper::modifyPrimaryKey(
-                    $this->logger,
-                    $this->clientWrapper->getBasicClient(),
-                    $config['destination'],
-                    $tableInfo['primaryKey'],
-                    $config['primary_key']
-                );
-            }
-            if (!empty($config['delete_where_column'])) {
-                // Delete rows
-                $deleteOptions = [
-                    'whereColumn' => $config['delete_where_column'],
-                    'whereOperator' => $config['delete_where_operator'],
-                    'whereValues' => $config['delete_where_values'],
-                ];
-                $this->clientWrapper->getBasicClient()->deleteTableRows($config['destination'], $deleteOptions);
-            }
-        } else {
-            $primaryKey = join(",", PrimaryKeyHelper::normalizeKeyArray($this->logger, $config['primary_key']));
-            $distributionKey = join(
-                ",",
-                PrimaryKeyHelper::normalizeKeyArray($this->logger, $config['distribution_key'])
-            );
-            if (!empty($config['columns'])) {
-                $this->createTable(
-                    $config['destination'],
-                    $config['columns'],
-                    $primaryKey,
-                    $distributionKey ?: null
-                );
-            } else {
-                try {
-                    $csvFile = new CsvFile($source, $config['delimiter'], $config['enclosure']);
-                    $header = $csvFile->getHeader();
-                } catch (Exception $e) {
-                    throw new InvalidOutputException('Failed to read file ' . $source . ' ' . $e->getMessage());
-                }
-                $this->createTable(
-                    $config['destination'],
-                    $header,
-                    $primaryKey,
-                    $distributionKey ?: null
-                );
-                unset($csvFile);
-            }
-            $this->metadataClient->postTableMetadata(
-                $config['destination'],
-                self::SYSTEM_METADATA_PROVIDER,
-                $this->getCreatedMetadata($systemMetadata)
-            );
-        }
-
-        $loadOptions = [
-            'delimiter' => $config['delimiter'],
-            'enclosure' => $config['enclosure'],
-            'columns' => !empty($config['columns']) ? $config['columns'] : [],
-            'incremental' => $config['incremental'],
-        ];
-        $tableQueue = $this->loadDataIntoTable($source, $config['destination'], $loadOptions);
-        $tableQueue->addMetadata(new MetadataDefinition(
-            $this->clientWrapper->getBasicClient(),
-            $config['destination'],
-            self::SYSTEM_METADATA_PROVIDER,
-            $this->getUpdatedMetadata($systemMetadata),
-            'table'
-        ));
-        return $tableQueue;
+        return $prefix . $filename;
     }
 
-    private function getBucketId($tableId)
-    {
-        $tableIdParts = $this->getTableIdParts($tableId);
-        return $tableIdParts[0] . '.' . $tableIdParts[1];
-    }
-
-    private function getTableIdParts($tableId)
-    {
-        return explode('.', $tableId);
-    }
-
-    private function createBucket($tableId, array $systemMetadata)
-    {
-        // Create bucket if not exists
-        $this->clientWrapper->getBasicClient()->createBucket($this->getBucketName($tableId), $this->getBucketStage($tableId));
-        $this->metadataClient->postBucketMetadata(
-            $this->getBucketId($tableId),
-            self::SYSTEM_METADATA_PROVIDER,
-            $this->getCreatedMetadata($systemMetadata)
-        );
-    }
-
-    private function getBucketName($tableId)
-    {
-        return substr($this->getTableIdParts($tableId)[1], 2);
-    }
-
-    private function getBucketStage($tableId)
-    {
-        return $this->getTableIdParts($tableId)[0];
-    }
-
-    /**
-     * @param array $systemMetadata
-     * @return array
-     */
-    private function getCreatedMetadata(array $systemMetadata)
-    {
-        $metadata[] = [
-            'key' => self::KBC_CREATED_BY_COMPONENT_ID,
-            'value' => $systemMetadata[self::SYSTEM_KEY_COMPONENT_ID],
-        ];
-        if (!empty($systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_CREATED_BY_CONFIGURATION_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID],
-            ];
-        }
-        if (!empty($systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_CREATED_BY_CONFIGURATION_ROW_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID],
-            ];
-        }
-        if (!empty($systemMetadata[self::SYSTEM_KEY_BRANCH_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_CREATED_BY_BRANCH_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_BRANCH_ID],
-            ];
-        }
-        return $metadata;
-    }
-
-    private function createTable($tableId, array $columns, $primaryKey, $distributionKey = null)
-    {
-        $tmp = new Temp();
-        $headerCsvFile = new CsvFile($tmp->createFile($this->getTableName($tableId) . '.header.csv'));
-        $headerCsvFile->writeRow($columns);
-        $options = ['primaryKey' => $primaryKey];
-        if (isset($distributionKey)) {
-            $options['distributionKey'] = $distributionKey;
-        }
-        $tableId = $this->clientWrapper->getBasicClient()->createTableAsync(
-            $this->getBucketId($tableId),
-            $this->getTableName($tableId),
-            $headerCsvFile,
-            $options
-        );
-        return $tableId;
-    }
-
-    private function getTableName($tableId)
-    {
-        return $this->getTableIdParts($tableId)[2];
-    }
-
-    private function loadDataIntoTable($sourcePath, $tableId, array $options)
-    {
-        if ($this->stagingStorageOutput === StrategyFactory::LOCAL) {
-            if (is_dir($sourcePath)) {
-                $fileId = $this->uploadSlicedFile($sourcePath);
-            } else {
-                $fileId = $this->clientWrapper->getBasicClient()->uploadFile(
-                    $sourcePath,
-                    (new FileUploadOptions())->setCompress(true)
-                );
-            }
-
-            $options['dataFileId'] = $fileId;
-            return new LoadTable($this->clientWrapper->getBasicClient(), $tableId, $options);
-        }
-
-        if ($this->stagingStorageOutput === StrategyFactory::WORKSPACE_ABS) {
-            $sourcePath = $this->specifySourceAbsPath($sourcePath);
-        }
-
-        $dataStorage = $this->strategy->getDataStorage();
-        $options = [
-            'dataWorkspaceId' => $dataStorage->getWorkspaceId(),
-            'dataObject' => $sourcePath,
-            'incremental' => $options['incremental'],
-            'columns' => $options['columns'],
-        ];
-
-        return new LoadTable($this->clientWrapper->getBasicClient(), $tableId, $options);
-    }
-
-    private function specifySourceAbsPath($sourcePath)
-    {
-        $path = $this->ensurePathDelimiter($this->sourcePath) . $sourcePath;
-        $absCredentials = $this->strategy->getDataStorage()->getCredentials();
-        $blobClient = BlobRestProxy::createBlobService($absCredentials['connectionString']);
-        try {
-            $options = new ListBlobsOptions();
-            $options->setPrefix($path);
-            $blobs = $blobClient->listBlobs($absCredentials['container'], $options);
-            $isSliced = false;
-            foreach ($blobs->getBlobs() as $blob) {
-                /* there can be multiple blobs with the same prefix (e.g `my`, `my-file`, ...), we're checking
-                    if there are blobs where the prefix is a directory. (e.g `my/` or `my-file/`) */
-                if (substr($blob->getName(), 0, strlen($path) + 1) === $path . '/') {
-                    $isSliced = true;
-                }
-            }
-            if ($isSliced) {
-                $path .= '/';
-            }
-        } catch (\Exception $e) {
-            throw new InvalidOutputException('Failed to list blobs ' . $e->getMessage(), 0, $e);
-        }
-        return $path;
-    }
-
-    protected function ensurePathDelimiter($path)
-    {
-        return $this->ensureNoPathDelimiter($path) . '/';
-    }
-
-    protected function ensureNoPathDelimiter($path)
-    {
-        return rtrim($path, '\\/');
-    }
-
-    /**
-     * Uploads a sliced table to storage api. Takes all files from the $source folder
-     *
-     * @param string $source Slices folder
-     * @return string
-     * @throws ClientException
-     */
-    private function uploadSlicedFile($source)
-    {
-        $finder = new Finder();
-        $slices = $finder->files()->in($source)->depth(0);
-        $sliceFiles = [];
-        /** @var SplFileInfo $slice */
-        foreach ($slices as $slice) {
-            $sliceFiles[] = $slice->getPathname();
-        }
-
-        // upload slices
-        $fileUploadOptions = new FileUploadOptions();
-        $fileUploadOptions
-            ->setIsSliced(true)
-            ->setFileName(basename($source))
-            ->setCompress(true);
-        return $this->clientWrapper->getBasicClient()->uploadSlicedFile($sliceFiles, $fileUploadOptions);
-    }
-
-    /**
-     * @param array $systemMetadata
-     * @return array
-     */
-    private function getUpdatedMetadata(array $systemMetadata)
-    {
-        $metadata[] = [
-            'key' => self::KBC_LAST_UPDATED_BY_COMPONENT_ID,
-            'value' => $systemMetadata[self::SYSTEM_KEY_COMPONENT_ID],
-        ];
-        if (!empty($systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_LAST_UPDATED_BY_CONFIGURATION_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID],
-            ];
-        }
-        if (!empty($systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_LAST_UPDATED_BY_CONFIGURATION_ROW_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID],
-            ];
-        }
-        if (!empty($systemMetadata[self::SYSTEM_KEY_BRANCH_ID])) {
-            $metadata[] = [
-                'key' => self::KBC_LAST_UPDATED_BY_BRANCH_ID,
-                'value' => $systemMetadata[self::SYSTEM_KEY_BRANCH_ID],
-            ];
-        }
-        return $metadata;
-    }
 }
